requirements.txt
Flask==3.0.3 
google-generativeai==0.8.3 
speechrecognition==3.10.4 
pyttsx3==2.98 
python-dotenv==1.0.1 
pyaudio==0.2.14
gTTS
speechrecognition==3.10.4 
openai==1.12.0
beautifulsoup4==4.12.3
requests==2.31.0